requirements.txt
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.3.3
azureml==0.2.7
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
Click==7.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
filelock==3.0.12
Flask==1.1.1
h5py==2.10.0
httplib2==0.9.2
idna==2.6
importlib-metadata==1.5.0
importlib-resources==1.0.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
netifaces==0.10.4
numpy==1.18.1
oauth==1.0.1
olefile==0.45.1
pandas==1.0.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.13.2
six==1.14.0
system-service==0.3
systemd-python==234
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.0.2
wadllib==1.3.2
wrapt==1.11.2
xkit==0.0.0
zipp==2.2.0
zope.interface==4.3.2
